10.5: improved casting.

+++ fuse_ipc.h
@@ -55,7 +55,7 @@ do {                                                   \
     (spc2) = (char *)(fiov)->base + (sizeof(*(spc1))); \
 } while (0)
 
-#define FU_AT_LEAST(siz) max((siz), 160)
+#define FU_AT_LEAST(siz) max((size_t)(siz), (size_t)160)
 
 struct fuse_ticket;
 struct fuse_data;
